+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, and this project adheres to Semantic Versioning.
 
+## [0.5.1]
+
+### Fixed
+- Language system not working with non-English characters. Parser now utilises UTF-8.
+
 ## [0.5.0]
 
 ### Added